importeBQ/requirements.txt
functions-framework==3.5.0
cloudevents==1.11.0
google.cloud
google-cloud-bigquery
google-cloud-storage
